moved it tests to a separate dir, and separated test tasks in grunt
