+++ .gitignore
@@ -1,6 +1,6 @@
 /target
 .idea/*
 *.iml
-app/node_modules/
+/node_modules/
 
 npm-debug.log
