+++ .travis.yml
@@ -1,3 +1,4 @@
 language: node_js
 node_js:
   - "0.12"
+before_install: npm install -g grunt-cli
